Add appkey
At least, this way, it's no harder to open a gdb session and print the
freakin' thing than "deobfuscating" the appkey.

And most of all, it makes it possible for potential contributors to
build the project without creating a separate key, which is highly
convenient.

But, yeah, this is silly.

Signed-off-by: Anton Lofgren <alofgren@op5.com>