audio: Present tracks as wave files, instead of raw pcm
By prepending a wav header to the pcm data, we've in effect a valid wave
file, perfectly playable in most, if not all, existing media players.
More importantly, it lets us avoid manually specifying track information
such as stereo/mono and sample rate, and lets us calculate the size of
the entire track, which in turn allows media players (such as mplayer)
to in its turn calculate the duration of the track.

Highly convenient!

This also renames the file previously known as pcm to wav, for obvious
reasons.

Signed-off-by: Anton Lofgren <anton.lofgren@gmail.com>