Disable FUSE entry caching
By default, FUSE caches entry names for 1 second. This is awesome for
disk-backed file systems; not so much for network-backed ones (such as
spotifile), since it means that a stat for an as of yet unknown name
results in that name still being unresolvable even when the entry has
been created, for example as the result of a readdir operation.

In particular, this causes great annoyance with the autocompleters that
many shells use, since it might trigger, for example, a full traversal
and subsequent loading of all playlists and their tracks while looking
for a completion for the file that really does exist.

Signed-off-by: Anton Lofgren <anton.lofgren@gmail.com>