Remove remember_me option
It wasn't ever read, and was added on a whim. That's a bad idea, always.
Remove it to get rid of the clutter. Maybe I'll add it back in the
future if I can come up with some non-contrived use case for it (that
can't be solved by a presistent settings file).

Signed-off-by: Anton Lofgren <anton.lofgren@gmail.com>