callbacks: Handle play_token_lost
We handle this the only way we can, as a passive data source; When
another application claims the token, we simply accept our fate and
unload the track. The libSpotify recommendation in this case is to
"behave just as if the user pressed the pause button".

Since we don't have a pause button, we'll have to make do with this
solution for now. If this proves troublesome in the future, one could
imagine some ioctl (or something else that requires active user input)
that reclaims the token. Automatically resuming playback is out of the
question (the libSpotify docs even forbid it explicitly:

"[the application] MUST NOT automatically resume playback but must
instead wait for the user to press play.")

Signed-off-by: Anton Lofgren <anton.lofgren@gmail.com>