Add slash-replacement
This patch changes the previous method of name sanitation by, instead of
removing slashes, replacing them with a similar character - the
'DIVISION SLASH'. This is a cosmetic change, but may cause problems if a
user is using a font that doesn't support it. If such issues show up in
the wild, it might be worth considering making the replacement character
configurable.

This fixes #9.

Signed-off-by: Anton Lofgren <anton.lofgren@gmail.com>