Add mount.spotifile wrapper
In order to be able to support the use case where spotifile is defined
in fstab, we need to conform to the calling convention for mount
executables and accept a device as a second argument. However, rather
than cluttering the main binary with this, we add a trivial wrapper
script which simply throws away the second argument and execs the main
binary.

The wrapper also discards options that are not supported by libfuse, but
are used by mount prior to executing the wrapper.

Signed-off-by: Anton Lofgren <anton.lofgren@gmail.com>