Disallow multiple concurrent listeners
By using a state-wide marker to indicate if music is currently playing,
we can avoid the problem of multiple processes reading track files at
once, which is not something I really care to support.

The current - rather naïve, but nonetheless efficient - approach is to
simply refuse open()ing a track.wav if another one is already opened,
until that file is released.

Signed-off-by: Anton Lofgren <anton.lofgren@gmail.com>