refactoring: Encapsulate entity ops
Since the number of valid ops for a given entity has grown, and I expect
it to continue to do so, it's a bit neater to pass around a struct
containing the supported operations instead of enumerating all of them.

This should make adding new operations a bit easier in the future.

Signed-off-by: Anton Lofgren <anton.lofgren@gmail.com>