An attempt at a sane playlist directory structure
Subdiring each playlist directly (with a pure music dir) turned out to
be suboptimal. This moves the music dir distinction one level up, in
order to explore a solution where we'll be able to point at *one*
directory and say "This is where the music is".

This is nice, because it allows us to do

~/spotifile/playlists/music$ mplayer -shuffle */*.wav

without loading a lot of superfluous metadata.

Signed-off-by: Anton Lofgren <anton.lofgren@gmail.com>