Add XPFS playlists to playlists/music/ directories
The XPFS format allows us to provide information such as duration for media
players that support it. This is helpful, since media players in general try to
read such information from the header of a track, which is not possible for us
to calculate (due to differing bitrates etc.) without doing a full unload/load
cycle for each track.

This fixes #37.

Signed-off-by: Anton Löfgren <anton.lofgren@gmail.com>